requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
opencv-python>=4.5.3
pillow>=8.3.1
streamlit>=1.22.0
requests>=2.26.0
tqdm>=4.62.0
matplotlib>=3.4.3
seaborn>=0.11.2
efficientnet>=1.0.0
albumentations>=1.0.3
gdown>=4.7.1
kaggle>=1.5.16

Pillow==9.5.0
openai==1.0.0
scikit-learn==1.2.2
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
requests==2.28.2
python-dotenv==1.0.0
streamlit==1.22.0

absl-py==2.1.0
altair==5.5.0
astunparse==1.6.3
attrs==25.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1

kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
narwhals==1.30.0
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
protobuf==3.20.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.1
referencing==0.36.2
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.43.2
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-probability==0.25.0
tensorflow-protobuf==2.11.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2

python-dotenv>=0.19.0
twilio>=7.0.0
